requirement.txt
google-auth==2.36.0
google-oauth2-tool==0.0.3
google-api-python-client==2.154.0
python-dotenv==1.0.1
langchain==0.2.17
langchain-community==0.2.19
pypdf==5.1.0
langchain-ollama==0.1.3